Remove unused code and classes
