+++ app/src/main/AndroidManifest.xml
@@ -10,8 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.CleanNotes">
-        <activity android:name=".presentation.ui.AddNoteActivity" />
-        <activity android:name=".presentation.ui.MainActivity">
+        <activity android:name=".presentation.view.activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
