+++ app/src/main/java/br/com/mmdevelopment/cleannotes/presentation/view/activity/MainActivity.kt
@@ -1,0 +1,32 @@
+package br.com.mmdevelopment.cleannotes.presentation.view.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.NavigationUI
+import br.com.mmdevelopment.cleannotes.R
+import br.com.mmdevelopment.cleannotes.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        NavigationUI.setupActionBarWithNavController(this, getNavController())
+    }
+
+    override fun onSupportNavigateUp(): Boolean {
+        return getNavController().navigateUp()
+    }
+
+    private fun getNavController(): NavController {
+        val navHostFragment = supportFragmentManager
+            .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        return navHostFragment.navController
+    }
+}
